* Moved class implementation into separate files * Cleaned up visual/collision, etc attachment mechanism * Added drawing of intertias for SDF
