+++ viz/OSGSegmentCallback.cpp
@@ -1,0 +1,6 @@
+#include "OSGSegmentCallback.h"
+
+OSGSegmentCallback::OSGSegmentCallback()
+{
+
+}
